move lib and ui to app
