+++ frontend/web/stories/lib/Header.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
-import { Header } from '../../lib/ui/Header/Header';
+import { Header } from '@/app/ui/Header/Header';
 
 const meta = {
   title: 'Example/Header',
